#53 #54 #61 #62 #63 #64 #65 Implementation and workin
#53 Basic integrated UI implemented; we are now able to move around the app without needing to change URLs

#54 Added a different menu based on the user role(AdminMenu.js, PatientMenu.js and DoctorMenu.js); also has a common menu that is common amoung every role that includes: / (to the landing page), /about that redirects to the about page (not implemented yet); also every menu includes the logout button

#61 Patient registration as user: PatientRegister.js; from the landing page the patient can register as a user if there is already a profile created by the admin with his email in the database

#62 Update the profile as a Patient: PatientProfile.js; implements the PatientUpdate and PatientInfo; displays the PatientProfile information and lets the Patient edit some of the fields: also sending a verification email if sensitive information is changed (implemented in ConfirmEmail.js)

#63 Deletes the Patient USER account (not the profile): implemented in RequestDeleteAccount.js and ConfirmDeleteAccount.js; Patient requests the account deletion in the app and then uses the link sent to his email to confirm the deletion

#64 Login feature implemented in Login.js: each user will be redirected to a page based on his role; if the user tries to access a page that he doesn't have permissions to access he will be redirected to the landing page (code implemented in PrivateRoute.js)

#65 Patient can login to the app and access his private information that is linked to his profile (PatientProfile.js)

Non mentioned USs features:

ResetPassword: User can request a reset password by giving his email: an email will be sent and if used he can provide a new password (RequestResetPassword.js, ResetPassword.js)
As an Admin.... : I can register, list, update and delete users (RegisterUser.js, DeleteUser.js, UpdateUser.js, UserList.js)

Previous code changes (backend):

Login: The login now implements a JWT Security Token system used the validate the correct user: once the user logins, it returns the token and its role, used in the frontend (stored in the localStorage) to validate if the user is logged in and what is his role. All the [Authorized] methods in the backend now need to have the 'Bearer' token passed in the header or they will return "401 Unauthorized"; also defined in the appsettings.json, this token expires after 60 minutes and a new login will be needed.

Logout: the logout is now handled in the frontend, NOT IN THE BACKEND; when logging out the jwt token and the role is removed from the localStorage; this also means that when using postman the logout functionally doesn't really work